Intital Implementation Code
Functions:
- UDP
- Data transfer
- Sequence Numbers
- Multiple packets

Missing Functions
- ACKs
- Timers
- Handle Loss
- Handle Duplicates

Proffesor Notes on Intitial Implementation Code -

You have a good start for send﻿﻿﻿﻿﻿﻿﻿ing multi-packet UDP data.  Key areas that are missing are currently a timer needs to be added to the sender so that if an ACK isn't received the packet is resent.  ACKs need to be added to know the packet is received.  Also while you have sequence numbers for your packets you need to add logic to your receiver so that a duplicate packet isn't delivered.